+++ src/html-url.c
@@ -185,7 +185,7 @@ init_interesting (void)
      matches the user's preferences as specified through --ignore-tags
      and --follow-tags.  */
 
-  int i;
+  size_t i;
   interesting_tags = make_nocase_string_hash_table (countof (known_tags));
 
   /* First, add all the tags we know hot to handle, mapped to their
@@ -358,7 +358,7 @@ tag_find_urls (int tagid, struct taginfo *tag, struct map_context *ctx)
   int i, attrind;
   int first = -1;
 
-  for (i = 0; i < countof (tag_url_attributes); i++)
+  for (i = 0; i < (int) countof (tag_url_attributes); i++)
     if (tag_url_attributes[i].tagid == tagid)
       {
         /* We've found the index of tag_url_attributes where the
